+++ src/legacy/core_plugins/data/public/index.ts
@@ -22,8 +22,6 @@
 // and handled by the platform itself in the setup method
 // of the ExpressionExectorService
 // @ts-ignore
-import { getInterpreter } from 'plugins/interpreter/interpreter';
-// @ts-ignore
 import { renderersRegistry } from 'plugins/interpreter/registries';
 import { ExpressionsService, ExpressionsSetup } from './expressions';
 import { SearchService, SearchSetup } from './search';
@@ -31,7 +29,7 @@ import { QueryService, QuerySetup } from './query';
 import { FilterService, FilterSetup } from './filter';
 import { IndexPatternsService, IndexPatternsSetup } from './index_patterns';
 
-class DataPlugin {
+export class DataPlugin {
   // Exposed services, sorted alphabetically
   private readonly expressions: ExpressionsService;
   private readonly filter: FilterService;
@@ -48,6 +46,9 @@ class DataPlugin {
   }
 
   public setup(): DataSetup {
+    // TODO: this is imported here to avoid circular imports.
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const { getInterpreter } = require('plugins/interpreter/interpreter');
     return {
       expressions: this.expressions.setup({
         interpreter: {
@@ -70,13 +71,6 @@ class DataPlugin {
     this.query.stop();
   }
 }
-
-/**
- * We export data here so that users importing from 'plugins/data'
- * will automatically receive the response value of the `setup` contract, mimicking
- * the data that will eventually be injected by the new platform.
- */
-export const data = new DataPlugin().setup();
 
 /** @public */
 export interface DataSetup {
