+++ src/legacy/core_plugins/kibana/public/dashboard/index.js
@@ -39,7 +39,7 @@ import { DashboardListing, EMPTY_FILTER } from './listing/dashboard_listing';
 import { uiModules } from 'ui/modules';
 import 'ui/capabilities/route_setup';
 
-import { data } from 'plugins/data';
+import { data } from 'plugins/data/setup';
 data.search.loadLegacyDirectives();
 data.filter.loadLegacyDirectives();
