+++ src/legacy/core_plugins/kibana/public/visualize/index.js
@@ -29,7 +29,7 @@ import { VisualizeConstants } from './visualize_constants';
 import { FeatureCatalogueRegistryProvider, FeatureCatalogueCategory } from 'ui/registry/feature_catalogue';
 import { getLandingBreadcrumbs, getWizardStep1Breadcrumbs } from './breadcrumbs';
 
-import { data } from 'plugins/data';
+import { data } from 'plugins/data/setup';
 data.search.loadLegacyDirectives();
 data.filter.loadLegacyDirectives();
