+++ src/legacy/core_plugins/metrics/public/components/panel_config/gauge.js
@@ -44,7 +44,7 @@ import {
 } from '@elastic/eui';
 import { injectI18n, FormattedMessage } from '@kbn/i18n/react';
 import { Storage } from 'ui/storage';
-import { data } from 'plugins/data';
+import { data } from 'plugins/data/setup';
 const { QueryBarInput } = data.query.ui;
 const localStorage = new Storage(window.localStorage);
 import { getDefaultQueryLanguage } from '../lib/get_default_query_language';
