+++ src/legacy/core_plugins/metrics/public/components/panel_config/metric.js
@@ -40,7 +40,7 @@ import {
 import { FormattedMessage } from '@kbn/i18n/react';
 
 import { Storage } from 'ui/storage';
-import { data } from 'plugins/data';
+import { data } from 'plugins/data/setup';
 import { getDefaultQueryLanguage } from '../lib/get_default_query_language';
 const { QueryBarInput } = data.query.ui;
 const localStorage = new Storage(window.localStorage);
