+++ src/legacy/core_plugins/metrics/public/components/series_config.js
@@ -36,7 +36,7 @@ import {
   EuiSpacer,
 } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n/react';
-import { data } from 'plugins/data';
+import { data } from 'plugins/data/setup';
 import { Storage } from 'ui/storage';
 import { getDefaultQueryLanguage } from './lib/get_default_query_language';
 const { QueryBarInput } = data.query.ui;
