+++ src/legacy/core_plugins/metrics/public/components/vis_types/table/config.js
@@ -40,7 +40,7 @@ import {
   EuiTitle,
 } from '@elastic/eui';
 import { FormattedMessage, injectI18n } from '@kbn/i18n/react';
-import { data } from 'plugins/data';
+import { data } from 'plugins/data/setup';
 import { Storage } from 'ui/storage';
 import { getDefaultQueryLanguage } from '../../lib/get_default_query_language';
 const { QueryBarInput } = data.query.ui;
