+++ src/legacy/ui/public/agg_types/controls/filter.tsx
@@ -28,7 +28,9 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { AggConfig } from 'ui/vis';
-import { Query, data } from 'plugins/data';
+// @ts-ignore
+import { data } from 'plugins/data/setup';
+import { Query } from 'plugins/data';
 
 const { QueryBarInput } = data.query.ui;
 
@@ -104,7 +106,7 @@ function FilterRow({
           query={value}
           indexPatterns={[agg.getIndexPattern()]}
           appName="filtersAgg"
-          onChange={query => onChangeValue(id, query, customLabel)}
+          onChange={(query: Query) => onChangeValue(id, query, customLabel)}
           disableAutoFocus={!autoFocus}
           data-test-subj={dataTestSubj}
           bubbleSubmitEvent={true}
